Fixed the findKingAttackPath method
Added verification that hidden attack is actually a valid attack in the classes Rook, Queen, and Bishop. Initially the only check that a hidden attack right past the king was valid was simply that the attack was on the board. Added check for these three classes that ensures that it's both on the board and a valid attack.